requirements.txt
# Telegram Bot
aiogram>=3.4.1
aiohttp>=3.9.0

# Web Framework
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
WTForms>=3.1.0

# Database - Tortoise-ORM
tortoise-orm>=0.20.0
aerich>=0.7.2

# Cache & Session
redis>=5.0.1

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utils
uvicorn>=0.24.0
gunicorn>=21.2.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
